refactor: sections 일급 컬렉션 객체로 변경
